Removed the pre-compiled marked library.
Removed marked.min.js and updated the documentation where to obtain an
up-to-date copy of the markdown library.

Added a configuration variable js_render_lib that points to the HTTP path of
marked.min.js. This allows to use a distribution-supplied version of marked.

Added a SQLite schema of the database layout.

Improved on the documentation.
The install command now also copies the SQLite schema file needed to create the
database.